+++ src/locales/__internal__/pl-pl.ts
@@ -1,5 +1,5 @@
-import Locale from "./locale";
-import { pl as plDateLocale } from "./date-fns-locales";
+import Locale from "../locale";
+import { pl as plDateLocale } from "../date-fns-locales";
 
 const isSingular = (count: string | number): boolean =>
   (typeof count === "string" ? parseInt(count) : count) === 1;
